Improve decimal quantity

+++ src/Server/esWMS.Application/Functions/WarehouseUnitItems/WarehouseUnitItemDto.cs
@@ -12,7 +12,7 @@ namespace esWMS.Application.Functions.WarehouseUnitItems
         public string ProductId { get; set; } = null!;
         public bool IsMediaOfWarehouseUnit { get; set; }
         public decimal Quantity { get; set; }
-        public int BlockedQuantity { get; set; }
+        public decimal BlockedQuantity { get; set; }
         public DateTime? BestBefore { get; set; }
         [DisplayName("Batch/Lot")]
         public string? BatchLot { get; set; }
